Apply changes from code review
