+++ frontend/src/core/components/DefaultLayout/AppNavbar/AppNavbar.scss
