+++ frontend/src/core/components/DefaultLayout/AppSidebar/AppSidebar.scss
